requirements.txt
# Core dependencies
pillow>=11.3.0
pyyaml>=6.0.2
opencv-python>=4.12.0.88
numpy>=2.2.6
requests>=2.32.5
tkinterdnd2>=0.3.0

# Video Processing
scenedetect>=0.6.7
ffmpeg-python>=0.2.0

# Face Detection
facexlib>=0.3.0
face_recognition>=1.3.0

# YOLOv5 dependencies
pandas>=2.3.2
torch>=2.8.0
torchvision>=0.23.0
ultralytics>=8.3.202

# Background Removal
onnxruntime-gpu>=1.22.0
onnxruntime>=1.22.1

# Image Enhancement and Upscaling
gfpgan>=1.3.8

# SDXL Inpainting dependencies
diffusers>=0.35.1
transformers>=4.56.2
accelerate>=1.10.1
huggingface_hub>=0.35.0
peft>=0.17.1

# Providers dependencies
instaloader>=4.14.2
snscrape>=0.7.0.20230622
praw>=7.8.1
google-images-download>=2.8.0
bing-image-downloader>=1.1.2
flickrapi>=2.4.0
pytube>=15.0.0

# System Monitoring
psutil>=7.1.0
nvidia-ml-py>=13.580.82
